refactor(noodle): extract public generation service
