+++ packages/server/src/services/noodle/noodle-public-support.ts
@@ -1,0 +1,193 @@
+import {
+  extractNoodleMentionHandles,
+  PROFESSOR_MARI_ID,
+  type NoodleAccount,
+  type NoodleAccountProfileSettings,
+  type NoodleBootstrap,
+  type NoodleInteractionType,
+  type NoodleSettings,
+} from "@marinara-engine/shared";
+import { createCharactersStorage } from "../storage/characters.storage.js";
+import { createNoodleStorage, parseNoodleAvatarCrop } from "../storage/noodle.storage.js";
+import { isNoodleProfileGenerated } from "./noodle-profile-selection.js";
+
+const PROFESSOR_MARI_NOODLE_BIO =
+  "She/Her | 18+ | Skill Issue | Your Assistant After Hours (hey, I get to do fun stuff, too!) | Simp for Il Dottore 24/7 | LLMs Fan";
+
+export function parseRecord(value: unknown): Record<string, unknown> {
+  if (!value) return {};
+  if (typeof value === "string") {
+    try {
+      const parsed = JSON.parse(value);
+      return parsed && typeof parsed === "object" && !Array.isArray(parsed) ? (parsed as Record<string, unknown>) : {};
+    } catch {
+      return {};
+    }
+  }
+  return typeof value === "object" && !Array.isArray(value) ? (value as Record<string, unknown>) : {};
+}
+
+export function characterAvatarCrop(row: { data: unknown }) {
+  return parseNoodleAvatarCrop(parseRecord(parseRecord(row.data).extensions).avatarCrop);
+}
+
+export function characterNameFromRow(row: { data: unknown } | null | undefined) {
+  const data = parseRecord(row?.data);
+  return typeof data.name === "string" && data.name.trim() ? data.name.trim() : "Character";
+}
+
+export function getErrorMessage(error: unknown): string {
+  return error instanceof Error ? error.message : String(error);
+}
+
+export function mentionedCharacterAccounts(accounts: NoodleAccount[], content: string): NoodleAccount[] {
+  const mentionedHandles = new Set(extractNoodleMentionHandles(content));
+  if (mentionedHandles.size === 0) return [];
+  return accounts.filter(
+    (account) => account.kind === "character" && mentionedHandles.has(account.handle.toLowerCase()),
+  );
+}
+
+export function mentionedAccountMetadata(accounts: NoodleAccount[]) {
+  return {
+    mentionedAccountIds: accounts.map((account) => account.id),
+    mentionedEntityIds: accounts.map((account) => account.entityId),
+  };
+}
+
+export function generatedProfileSettings(location: string, bannerUrl: string | null): NoodleAccountProfileSettings {
+  return {
+    profileGenerated: true,
+    location,
+    bannerUrl: bannerUrl ?? "",
+  };
+}
+
+export async function ensureProfessorMariAccount(
+  noodle: ReturnType<typeof createNoodleStorage>,
+  characters: ReturnType<typeof createCharactersStorage>,
+) {
+  const row = await characters.getById(PROFESSOR_MARI_ID);
+  const account = await noodle.upsertAccountFromProfile({
+    kind: "character",
+    entityId: PROFESSOR_MARI_ID,
+    displayName: row ? characterNameFromRow(row) : "Professor Mari",
+    avatarUrl: row?.avatarPath ?? "/sprites/mari/Mari_profile.png",
+    avatarCrop: row ? characterAvatarCrop(row) : null,
+    bio: PROFESSOR_MARI_NOODLE_BIO,
+    invited: true,
+    syncIdentity: true,
+  });
+  if (
+    account.settings.profile.profileManuallyEdited !== true &&
+    (account.bio !== PROFESSOR_MARI_NOODLE_BIO ||
+      !isNoodleProfileGenerated(account) ||
+      !account.settings.profile.location)
+  ) {
+    await noodle.updateAccountProfile(account.id, {
+      handle: account.handle || "professor_mari",
+      displayName: account.displayName || "Professor Mari",
+      bio: PROFESSOR_MARI_NOODLE_BIO,
+      avatarUrl: account.avatarUrl || row?.avatarPath || "/sprites/mari/Mari_profile.png",
+      profile: generatedProfileSettings("Marinara Engine", null),
+    });
+  }
+}
+
+export async function ensurePersonaAccounts(
+  noodle: ReturnType<typeof createNoodleStorage>,
+  characters: ReturnType<typeof createCharactersStorage>,
+) {
+  const personas = await characters.listPersonas();
+  const livePersonaIds = new Set<string>();
+  for (const persona of personas) {
+    livePersonaIds.add(persona.id);
+    await noodle.upsertAccountFromProfile({
+      kind: "persona",
+      entityId: persona.id,
+      displayName: persona.convoDisplayName || persona.name || "User",
+      avatarUrl: persona.avatarPath ?? null,
+      avatarCrop: parseNoodleAvatarCrop(persona.avatarCrop),
+      bio: persona.aboutMe || persona.description || "",
+      invited: true,
+    });
+  }
+  return livePersonaIds;
+}
+
+function filterStalePersonaAccounts(bootstrap: NoodleBootstrap, livePersonaIds: Set<string>): NoodleBootstrap {
+  return {
+    ...bootstrap,
+    accounts: bootstrap.accounts.filter(
+      (account) => account.kind !== "persona" || livePersonaIds.has(account.entityId),
+    ),
+  };
+}
+
+function filterExcludedNoodleAccounts(bootstrap: NoodleBootstrap, settings: NoodleSettings): NoodleBootstrap {
+  if (settings.allowProfessorMari) return bootstrap;
+  return {
+    ...bootstrap,
+    accounts: bootstrap.accounts.filter(
+      (account) => account.kind !== "character" || account.entityId !== PROFESSOR_MARI_ID,
+    ),
+  };
+}
+
+export async function bootstrapVisibleNoodle(
+  noodle: ReturnType<typeof createNoodleStorage>,
+  characters: ReturnType<typeof createCharactersStorage>,
+) {
+  const settings = await noodle.getSettings();
+  const livePersonaIds = await ensurePersonaAccounts(noodle, characters);
+  if (settings.allowProfessorMari) await ensureProfessorMariAccount(noodle, characters);
+  const existingCharacterAccounts = (await noodle.listAccounts()).filter(
+    (account) => account.kind === "character" && account.entityId !== PROFESSOR_MARI_ID,
+  );
+  const characterRowsById = new Map((await characters.list()).map((row) => [row.id, row]));
+  for (const account of existingCharacterAccounts) {
+    const row = characterRowsById.get(account.entityId);
+    if (!row) continue;
+    await noodle.upsertAccountFromProfile({
+      kind: "character",
+      entityId: row.id,
+      displayName: characterNameFromRow(row),
+      avatarUrl: row.avatarPath ?? null,
+      avatarCrop: characterAvatarCrop(row),
+      syncIdentity: true,
+    });
+  }
+  return filterExcludedNoodleAccounts(filterStalePersonaAccounts(await noodle.bootstrap(), livePersonaIds), settings);
+}
+
+export async function resolvePersonaAccount(
+  noodle: ReturnType<typeof createNoodleStorage>,
+  characters: ReturnType<typeof createCharactersStorage>,
+  personaId?: string,
+) {
+  const personas = await characters.listPersonas();
+  const persona =
+    personas.find((p) => p.id === personaId) ?? personas.find((p) => p.isActive === "true") ?? personas[0];
+  if (!persona) return null;
+  return noodle.upsertAccountFromProfile({
+    kind: "persona",
+    entityId: persona.id,
+    displayName: persona.convoDisplayName || persona.name || "User",
+    avatarUrl: persona.avatarPath ?? null,
+    avatarCrop: parseNoodleAvatarCrop(persona.avatarCrop),
+    bio: persona.aboutMe || persona.description || "",
+    invited: true,
+  });
+}
+
+export function interactionDigestVerb(type: NoodleInteractionType) {
+  if (type === "reply") return "replied on";
+  if (type === "repost") return "reposted";
+  if (type === "vote") return "voted in";
+  return "liked";
+}
+
+export function noodleDigestAccountLabel(account: Pick<NoodleAccount, "kind" | "displayName" | "handle">) {
+  const identity = `${account.displayName} (@${account.handle})`;
+  return account.kind === "persona" ? `Persona ${identity}` : identity;
+}
